requirement.txt
inputimeout==1.0.4
numpy==1.24.4
opencv-python==4.8.0.76
PyYAML==6.0.1
